// src/shared/types/FollowUp.ts
export interface FollowUp {
	totalExperiences: number;
	experiencesNaciente: number;
	experiencesCreciente: number;
	experiencesInspiradora: number;
	totalExperiencesRegistradas: number;
	totalExperiencesCreadas: number;
	totalInstitutionsWithExperiences: number;
	totalTeachersRegistered: number;
	totalExperiencesWithComments: number;
	totalExperiencesTestsKnow: number;
}